lab2

+++ tasks/mityaeva_d_striped_horizontal_matrix_vector/mpi/src/ops_mpi.cpp
@@ -2,6 +2,7 @@
 
 #include <mpi.h>
 
+#include <cstddef>
 #include <vector>
 
 #include "mityaeva_d_striped_horizontal_matrix_vector/common/include/common.hpp"
@@ -53,6 +54,7 @@ bool StripedHorizontalMatrixVectorMPI::ValidationImpl() {
 }
 
 bool StripedHorizontalMatrixVectorMPI::PreProcessingImpl() {
+  // Получаем размеры матрицы
   if (rank_ == 0) {
     n_ = static_cast<int>(GetInput()[0]);
     m_ = static_cast<int>(GetInput()[1]);
@@ -61,6 +63,7 @@ bool StripedHorizontalMatrixVectorMPI::PreProcessingImpl() {
   MPI_Bcast(&n_, 1, MPI_INT, 0, MPI_COMM_WORLD);
   MPI_Bcast(&m_, 1, MPI_INT, 0, MPI_COMM_WORLD);
 
+  // Рассылаем вектор b всем процессам
   full_vector_.resize(m_);
   if (rank_ == 0) {
     size_t vector_start_index = 2 + n_ * m_;
@@ -70,14 +73,17 @@ bool StripedHorizontalMatrixVectorMPI::PreProcessingImpl() {
   }
   MPI_Bcast(full_vector_.data(), m_, MPI_DOUBLE, 0, MPI_COMM_WORLD);
 
+  // Распределяем строки матрицы по процессам
   int rows_per_process = n_ / world_size_;
   int remainder = n_ % world_size_;
 
+  // Определяем количество строк для текущего процесса
   int local_row_count = rows_per_process;
   if (rank_ < remainder) {
     local_row_count++;
   }
 
+  // Определяем начальную строку для текущего процесса
   int start_row = 0;
   for (int i = 0; i < rank_; ++i) {
     int proc_rows = rows_per_process;
@@ -87,16 +93,20 @@ bool StripedHorizontalMatrixVectorMPI::PreProcessingImpl() {
     start_row += proc_rows;
   }
 
+  // Выделяем память для локальных строк
   local_rows_.resize(local_row_count, std::vector<double>(m_));
   local_result_.resize(local_row_count, 0.0);
 
+  // Процесс 0 распределяет данные
   if (rank_ == 0) {
+    // Процесс 0 копирует свои строки
     for (int i = 0; i < local_row_count; ++i) {
       for (int j = 0; j < m_; ++j) {
         local_rows_[i][j] = GetInput()[2 + (start_row + i) * m_ + j];
       }
     }
 
+    // Отправляем строки другим процессам
     for (int dest_rank = 1; dest_rank < world_size_; ++dest_rank) {
       int dest_row_count = rows_per_process;
       if (dest_rank < remainder) {
@@ -110,6 +120,8 @@ bool StripedHorizontalMatrixVectorMPI::PreProcessingImpl() {
         }
         dest_start_row += proc_rows;
       }
+
+      // Отправляем все строки для dest_rank одним сообщением
       std::vector<double> dest_rows_data(dest_row_count * m_);
       for (int i = 0; i < dest_row_count; ++i) {
         for (int j = 0; j < m_; ++j) {
@@ -119,18 +131,23 @@ bool StripedHorizontalMatrixVectorMPI::PreProcessingImpl() {
       MPI_Send(dest_rows_data.data(), dest_row_count * m_, MPI_DOUBLE, dest_rank, 0, MPI_COMM_WORLD);
     }
   } else {
+    // Другие процессы получают свои строки
     std::vector<double> received_data(local_row_count * m_);
     MPI_Recv(received_data.data(), local_row_count * m_, MPI_DOUBLE, 0, 0, MPI_COMM_WORLD, MPI_STATUS_IGNORE);
 
+    // Распаковываем данные в локальные строки
     for (int i = 0; i < local_row_count; ++i) {
       for (int j = 0; j < m_; ++j) {
         local_rows_[i][j] = received_data[i * m_ + j];
       }
     }
   }
+
+  // Инициализируем выходной вектор ТОЛЬКО на процессе 0
   if (rank_ == 0) {
     GetOutput().resize(n_, 0.0);
   } else {
+    // На других процессах очищаем выход
     GetOutput().clear();
   }
 
@@ -138,7 +155,8 @@ bool StripedHorizontalMatrixVectorMPI::PreProcessingImpl() {
 }
 
 bool StripedHorizontalMatrixVectorMPI::RunImpl() {
-  for (int i = 0; i < local_rows_.size(); ++i) {
+  // Каждый процесс вычисляет свою часть результата
+  for (size_t i = 0; i < local_rows_.size(); ++i) {  // Исправлено: int -> size_t
     double sum = 0.0;
     for (int j = 0; j < m_; ++j) {
       sum += local_rows_[i][j] * full_vector_[j];
@@ -146,6 +164,7 @@ bool StripedHorizontalMatrixVectorMPI::RunImpl() {
     local_result_[i] = sum;
   }
 
+  // Собираем результаты на процессе 0
   std::vector<int> recv_counts(world_size_);
   std::vector<int> displacements(world_size_);
 
@@ -162,6 +181,7 @@ bool StripedHorizontalMatrixVectorMPI::RunImpl() {
     displacement += recv_counts[i];
   }
 
+  // Собираем результаты только если есть что собирать
   if (n_ > 0) {
     MPI_Gatherv(local_result_.data(), local_result_.size(), MPI_DOUBLE, GetOutput().data(), recv_counts.data(),
                 displacements.data(), MPI_DOUBLE, 0, MPI_COMM_WORLD);
@@ -172,10 +192,14 @@ bool StripedHorizontalMatrixVectorMPI::RunImpl() {
 }
 
 bool StripedHorizontalMatrixVectorMPI::PostProcessingImpl() {
+  // В MPI версии только процесс 0 имеет выходные данные
+  // Другие процессы не должны иметь выхода, и это нормально
   if (rank_ == 0) {
     return !GetOutput().empty() && GetOutput().size() == static_cast<size_t>(n_);
   }
 
+  // Для процессов кроме 0, выход должен быть пустым
+  // Если по какой-то причине есть данные, очищаем их
   if (!GetOutput().empty()) {
     GetOutput().clear();
   }
